src/requirements.txt
aiohttp==3.11.14
asyncpraw==7.8.1
beautifulsoup4==4.13.3
disnake==2.10.1
geopy==2.4.1
haversine==2.9.0
humanize==4.12.2
motor==3.7.0
numpy==2.2.4
Pillow==11.1.0
psutil==7.0.0
pyfiglet==1.0.2
pymongo==4.11.3
pyrate_limiter<4.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
Requests==2.32.3
requests_cache==1.2.1
requests_ratelimiter==0.7.0
webcolors==24.11.1
yfinance==0.2.55
zalgo_text==0.6
